requirements.txt
google-analytics-data==0.18.15
google-auth==2.34.0

